requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.32.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
window==0.0.3
